Outsourced convolution calculation.
Function 'convCalc' can be called in two different ways: convCalc(convCalcObject) or convCalc(inputTag, outputTag, convMatrix). In both cases, the function convCalc extracts the required data from the mdtsObject, passes it to the (new) function 'convCalcFct' - which is not a method of this class or of any class, but a separate function in a '.m' file! - and expands the data of the mdtsObject by the return of the 'convCalcFct'.