Implemented cell array of convCalc objects.
Cell arrays of convCalc objects can be used in the function 'convCalc'. The function will execute all convolutions of the given convCalc objects sequentially and save the results in the data of the mdts object. Thus, convCalc objects in a cell array of convCalc objects can use the results of preceding convCalc objects as inputs.