Added function 'addEvent'.
Function expects name of the event and the point in time of the event as input. The point in time can be of type datenum or datetime. It can also be passed as index of the time array if an additional flag is set to true as thrid input argument.